requirements.txt
openai
PyPDF2
python-dotenv
pytest
slack_sdk

